refactor: interfaces package 신규 생성
interfaces package 신규 생성에 파일 변경
