--- src/main/java/io/mwkwon/lotto/domain/LottoGenerator.java
@@ -1,5 +1,0 @@
-package io.mwkwon.lotto.domain;
-
-public interface LottoGenerator {
-    BuyLottos createAutoLottos(LottoPayment lottoPayment);
-}
