+++ src/main/java/io/mwkwon/lotto/domain/LottoMachine.java
@@ -1,6 +1,7 @@
 package io.mwkwon.lotto.domain;
 
 import io.mwkwon.lotto.constant.LottoConstants;
+import io.mwkwon.lotto.interfaces.LottoGenerator;
 
 import java.util.ArrayList;
 import java.util.Collections;
