+++ src/main/java/io/mwkwon/lotto/interfaces/LottoGenerator.java
@@ -1,0 +1,8 @@
+package io.mwkwon.lotto.interfaces;
+
+import io.mwkwon.lotto.domain.BuyLottos;
+import io.mwkwon.lotto.domain.LottoPayment;
+
+public interface LottoGenerator {
+    BuyLottos createAutoLottos(LottoPayment lottoPayment);
+}
